requirements.txt
# Core Adaptive Speculative Decoding Dependencies
# This file contains the minimum required dependencies for basic functionality

# ========== Core Machine Learning Libraries ==========
torch>=2.5.0,<3.0.0
transformers>=4.40.0,<5.0.0
accelerate>=0.35.0

# ========== Inference and Serving ==========
vllm>=0.8.3
bitsandbytes>=0.45.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0,<3.0.0

# ========== Configuration Management ==========
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# ========== Data Processing and Analysis ==========
numpy>=1.26.0,<2.0.0
scipy>=1.14.0
pandas>=2.2.0
scikit-learn>=1.6.0
lightgbm>=4.6.0

# ========== Quality Evaluation ==========
evaluate>=0.4.0
datasets>=3.0.0
nltk>=3.9
rouge-score>=0.1.2
sacrebleu>=2.4.0
bert-score>=0.3.13

# ========== Visualization ==========
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.14.0

# ========== Monitoring and Logging ==========
wandb>=0.18.0
tensorboard>=2.18.0
tqdm>=4.66.0
psutil>=5.9.0

# ========== Async and Concurrency ==========
asyncio-mqtt>=0.16.0
aiofiles>=23.0.0
async-timeout>=4.0.0

# ========== Utilities ==========
click>=8.1.0
rich>=13.0.0
typer>=0.12.0
pathlib2>=2.3.0

# ========== File Handling ==========
aiofiles>=23.0.0
h5py>=3.9.0
pillow>=10.0.0

# ========== Network and HTTP ==========
requests>=2.32.0
aiohttp>=3.12.0
httpx>=0.27.0

# ========== Database (Optional) ==========
sqlalchemy>=2.0.0
redis>=5.0.0

# ========== Security ==========
cryptography>=42.0.0
python-multipart>=0.0.9

# ========== Type Checking Support ==========
typing-extensions>=4.8.0

# ========== Production Utilities ==========
gunicorn>=21.0.0
prometheus-client>=0.17.0